+++ src/engine/search/move_picker.rs
@@ -3,10 +3,8 @@ use crate::chess::movegen;
 use crate::chess::movegen::MovegenCache;
 use crate::chess::movelist::MoveList;
 use crate::chess::moves::Move;
-use crate::engine::eval::Eval;
 use crate::engine::search::move_ordering::score_move;
-use crate::engine::search::{PersistentState, SearchState};
-use crate::engine::see::see;
+use crate::engine::search::{move_ordering, PersistentState, SearchState};
 
 const MAX_MOVES: usize = u8::MAX as usize;
 
@@ -14,10 +12,11 @@ const MAX_MOVES: usize = u8::MAX as usize;
 enum GenStage {
     BestMove,
     GenCaptures,
-    Captures,
+    GoodCaptures,
     GenQuiets,
     Killer1,
     Killer2,
+    BadCaptures,
     ScoreQuiets,
     Quiets,
     Done,
@@ -33,6 +32,8 @@ pub struct MovePicker {
     stage: GenStage,
     idx: usize,
     captures_end: usize,
+    first_bad_capture: Option<usize>,
+    first_quiet: usize,
 }
 
 impl MovePicker {
@@ -47,6 +48,8 @@ impl MovePicker {
             stage: GenStage::BestMove,
             idx: 0,
             captures_end: 0,
+            first_bad_capture: None,
+            first_quiet: 0,
         }
     }
 
@@ -61,6 +64,8 @@ impl MovePicker {
             stage: GenStage::BestMove,
             idx: 0,
             captures_end: 0,
+            first_bad_capture: None,
+            first_quiet: 0,
         }
     }
 
@@ -71,49 +76,68 @@ impl MovePicker {
         state: &SearchState,
         plies: u8,
     ) -> Option<Move> {
-        if self.stage == GenStage::BestMove {
-            self.stage = GenStage::GenCaptures;
+        use GenStage::*;
+
+        if self.stage == BestMove {
+            self.stage = GenCaptures;
 
             if let Some(previous_best_move) = self.previous_best_move {
                 return Some(previous_best_move);
             }
         }
 
-        if self.stage == GenStage::GenCaptures {
-            self.stage = GenStage::Captures;
+        if self.stage == GenCaptures {
+            self.stage = GoodCaptures;
 
             movegen::generate_captures(game, &mut self.moves, &mut self.movegencache);
 
             self.captures_end = self.moves.len();
+            self.first_quiet = self.moves.len();
             self.score_moves(0, self.captures_end, game, persistent_state);
         }
 
-        if self.stage == GenStage::Captures {
-            if let Some(mv) = self.next_best_move(game, self.captures_end, self.only_captures) {
-                return Some(mv);
+        if self.stage == GoodCaptures {
+            if let Some((mv, score)) = self.next_best_move(self.captures_end) {
+                // If the move we just picked was a losing capture, we're going to skip the rest of the captures.
+                // Record that, and skip the remainder of the captures since we'll be trying quiet moves next.
+                if score < move_ordering::GOOD_CAPTURE_SCORE {
+                    self.first_bad_capture = Some(self.idx - 1);
+                    self.idx = self.captures_end;
+                } else {
+                    return Some(mv);
+                }
             }
 
-            self.stage = if self.only_captures {
-                GenStage::Done
+            if self.only_captures {
+                match self.first_bad_capture {
+                    // If we didn't see any bad captures before, we can skip straight to the end
+                    None => self.stage = Done,
+
+                    // If we saw any bad captures, go back and try those too
+                    Some(first_bad_capture_idx) => {
+                        self.idx = first_bad_capture_idx;
+                        self.stage = BadCaptures;
+                    }
+                }
             } else {
-                GenStage::GenQuiets
+                self.stage = GenQuiets;
             };
         }
 
-        if self.stage == GenStage::GenQuiets {
-            self.stage = GenStage::Killer1;
+        if self.stage == GenQuiets {
+            self.stage = Killer1;
 
             movegen::generate_quiets(game, &mut self.moves, &self.movegencache);
         }
 
-        if self.stage == GenStage::Killer1 {
-            self.stage = GenStage::Killer2;
+        if self.stage == Killer1 {
+            self.stage = Killer2;
 
             if let Some(killer1) = state.killer_moves.get_0(plies) {
-                for i in self.idx..self.moves.len() {
+                for i in self.first_quiet..self.moves.len() {
                     if self.moves.get(i) == killer1 {
-                        self.moves.swap(self.idx, i);
-                        self.idx += 1;
+                        self.moves.swap(self.first_quiet, i);
+                        self.first_quiet += 1;
 
                         if Some(killer1) != self.previous_best_move {
                             return Some(killer1);
@@ -123,14 +147,23 @@ impl MovePicker {
             }
         }
 
-        if self.stage == GenStage::Killer2 {
-            self.stage = GenStage::ScoreQuiets;
+        if self.stage == Killer2 {
+            match self.first_bad_capture {
+                // If we didn't see any bad captures before, we can skip straight to the end
+                None => self.stage = ScoreQuiets,
+
+                // If we saw any bad captures, go back and try those too
+                Some(first_bad_capture_idx) => {
+                    self.idx = first_bad_capture_idx;
+                    self.stage = BadCaptures;
+                }
+            }
 
             if let Some(killer2) = state.killer_moves.get_1(plies) {
-                for i in self.idx..self.moves.len() {
+                for i in self.first_quiet..self.moves.len() {
                     if self.moves.get(i) == killer2 {
-                        self.moves.swap(self.idx, i);
-                        self.idx += 1;
+                        self.moves.swap(self.first_quiet, i);
+                        self.first_quiet += 1;
 
                         if Some(killer2) != self.previous_best_move {
                             return Some(killer2);
@@ -140,32 +173,40 @@ impl MovePicker {
             }
         }
 
-        if self.stage == GenStage::ScoreQuiets {
-            self.stage = GenStage::Quiets;
-            self.score_moves(self.idx, self.moves.len(), game, persistent_state);
-        }
-
-        if self.stage == GenStage::Quiets {
-            if let Some(mv) = self.next_best_move(game, self.moves.len(), false) {
+        if self.stage == BadCaptures {
+            if let Some((mv, _)) = self.next_best_move(self.captures_end) {
                 return Some(mv);
             }
 
-            self.stage = GenStage::Done;
+            self.stage = if self.only_captures {
+                Done
+            } else {
+                ScoreQuiets
+            };
         }
 
-        if self.stage == GenStage::Done {
+        if self.stage == ScoreQuiets {
+            self.stage = Quiets;
+            self.idx = self.first_quiet;
+            self.score_moves(self.idx, self.moves.len(), game, persistent_state);
+        }
+
+        if self.stage == Quiets {
+            if let Some((mv, _)) = self.next_best_move(self.moves.len()) {
+                return Some(mv);
+            }
+
+            self.stage = Done;
+        }
+
+        if self.stage == Done {
             return None;
         }
 
         unreachable!()
     }
 
-    fn next_best_move(
-        &mut self,
-        game: &Game,
-        limit: usize,
-        skip_bad_captures: bool,
-    ) -> Option<Move> {
+    fn next_best_move(&mut self, limit: usize) -> Option<(Move, i32)> {
         loop {
             if self.idx == limit {
                 return None;
@@ -199,11 +240,7 @@ impl MovePicker {
                 continue;
             }
 
-            if skip_bad_captures && !see(game, best_move, Eval(1)) {
-                continue;
-            }
-
-            return Some(best_move);
+            return Some((best_move, best_move_score));
         }
     }
 
@@ -325,5 +362,27 @@ mod tests {
         }
 
         assert_eq!(moves.len(), 1);
+    }
+
+    #[test]
+    fn test_movepicker_bug_after_see_move_ordering_1() {
+        crate::init();
+
+        let game = Game::from_fen("r2k3r/1b4bq/8/3R4/8/8/7B/4K2R b K - 3 2").unwrap();
+
+        let mut moves: Vec<Move> = Vec::new();
+        let mut move_provider = MovePicker::new(Some(Move::new(D8, E7)));
+
+        let mut search_state = SearchState::new();
+        let persistent_state = PersistentState::new();
+
+        search_state.killer_moves.try_push(0, Move::new(B7, D5));
+        search_state.killer_moves.try_push(0, Move::new(D8, E8));
+
+        while let Some(m) = move_provider.next(&game, &persistent_state, &search_state, 0) {
+            moves.push(m);
+        }
+
+        assert_eq!(moves.len(), 4);
     }
 }